requirements.txt
dnspython==2.3.0
colorama